rename debug dir to dbg
