+++ main/src/ca/uwaterloo/flix/language/dbg/FormatSourceLocation.scala
@@ -1,4 +1,4 @@
-package ca.uwaterloo.flix.language.debug
+package ca.uwaterloo.flix.language.dbg
 
 import ca.uwaterloo.flix.language.ast.SourceLocation
