+++ main/src/ca/uwaterloo/flix/language/errors/StratificationError.scala
@@ -18,7 +18,7 @@ package ca.uwaterloo.flix.language.errors
 
 import ca.uwaterloo.flix.language.CompilationMessage
 import ca.uwaterloo.flix.language.ast._
-import ca.uwaterloo.flix.language.debug.{Audience, FormatType}
+import ca.uwaterloo.flix.language.dbg.{Audience, FormatType}
 import ca.uwaterloo.flix.util.Formatter
 
 /**
