+++ main/src/ca/uwaterloo/flix/language/phase/unification/BoolUnification.scala
@@ -19,7 +19,7 @@ import ca.uwaterloo.flix.api.Flix
 import ca.uwaterloo.flix.language.ast.Scheme.InstantiateMode
 import ca.uwaterloo.flix.language.ast.Type.eraseAliases
 import ca.uwaterloo.flix.language.ast._
-import ca.uwaterloo.flix.language.debug.Audience
+import ca.uwaterloo.flix.language.dbg.Audience
 import ca.uwaterloo.flix.util.Result.{Err, Ok}
 import ca.uwaterloo.flix.util.{InternalCompilerException, Result}
