+++ main/src/ca/uwaterloo/flix/runtime/shell/Shell.scala
@@ -18,7 +18,7 @@ package ca.uwaterloo.flix.runtime.shell
 
 import ca.uwaterloo.flix.api.{Flix, Version}
 import ca.uwaterloo.flix.language.ast.Ast
-import ca.uwaterloo.flix.language.debug.Audience
+import ca.uwaterloo.flix.language.dbg.Audience
 import ca.uwaterloo.flix.runtime.CompilationResult
 import ca.uwaterloo.flix.util.Formatter.AnsiTerminalFormatter
 import ca.uwaterloo.flix.util._
