+++ main/test/ca/uwaterloo/flix/language/LanguageSuite.scala
@@ -16,7 +16,7 @@
 
 package ca.uwaterloo.flix.language
 
-import ca.uwaterloo.flix.language.debug.TestFormatType
+import ca.uwaterloo.flix.language.dbg.TestFormatType
 import ca.uwaterloo.flix.language.phase.PhaseSuite
 import org.scalatest.Suites
